capstone/smart-campus-security/requirements.txt
# Core Dependencies
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.0.0

# Face Recognition
# insightface==0.7.3  # Requires C++ Build Tools, skipping for now
# onnxruntime==1.16.0
deepface==0.0.92
tensorflow-cpu==2.15.0  # Use CPU version for stability on laptops/Pi
tf-keras==2.15.0

# Object Detection
ultralytics==8.0.200
# TFLite runtime (optional - for optimized inference)
# tflite-runtime can be installed separately if needed

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Dashboard
streamlit==1.28.1
plotly==5.17.0
streamlit-webrtc==0.47.1

# Database & Data Processing
pandas==2.0.3
sqlalchemy==2.0.23

# Machine Learning
scikit-learn==1.3.2
prophet==1.1.5

# Alerts
pygame==2.5.2

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Image Processing
imutils==0.5.4

# Logging
loguru==0.7.2

# Export
reportlab==4.0.7
fpdf2==2.7.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional utilities
tqdm==4.66.1
scipy==1.11.3
